Move WatchStateUpdater to watch module
